Create mainApplication
Ok this isn't finished but I thought I'd post it anyway I can work on it more but I work late tonight so I have to stop for today. Still need to finish the insert for Practitioners and write a print function. I'm also kind of confused about whether we're prompting for usernames, passwords and databases or if we're going hard-code all of ours or just one? Like I said kind of confused. Let me know if my method doesn't make any sense, I probably over-complicated it but it's what I have so far.